refactor: extracted export functionality to controller
