+++ server/src/controllers/templates/reporter.ts
@@ -1,4 +1,7 @@
-export default function (client: string, main: string): string {
+import { IIntegration } from '@lumieducation/h5p-server';
+import fsExtra from 'fs-extra';
+
+function createReporter(client: string, main: string): string {
     return `<script>
     !(function (e) {
         function t(t) {
@@ -108,4 +111,40 @@ export default function (client: string, main: string): string {
 </script>
 <script>${client}</script>
 <script>${main}</script>`;
+}
+
+export default function (
+    integration: IIntegration,
+    scriptsBundle: string,
+    stylesBundle: string,
+    contentId: string
+): string {
+    const reporterClient = fsExtra.readFileSync(
+        `${__dirname}/../../../reporter-client/build/static/js/2.chunk.js`,
+        {
+            encoding: 'utf-8'
+        }
+    );
+
+    const reporterMain = fsExtra.readFileSync(
+        `${__dirname}/../../../reporter-client/build/static/js/main.chunk.js`,
+        {
+            encoding: 'utf-8'
+        }
+    );
+
+    return `<!doctype html>
+<html class="h5p-iframe">
+<head>
+<meta charset="utf-8">                    
+<script>H5PIntegration = ${JSON.stringify(integration)};
+${scriptsBundle}</script>
+<style>${stylesBundle}</style>
+</head>
+<body>
+<div id="root"></div>
+${createReporter(reporterClient, reporterMain)}
+<div style="margin: 20px auto; padding: 20px;  max-width: 840px; box-shadow: 0px 2px 1px -1px rgba(0,0,0,0.2), 0px 1px 1px 0px rgba(0,0,0,0.14), 0px 1px 3px 0px rgba(0,0,0,0.12)" class="h5p-content lag" data-content-id="${contentId}"></div>                
+</body>
+</html>`;
 }
